debug: add option for overwriting database file
When we want to add an attribute to a model (i.e. a new column to a table), the import functionnality does not work because this new column is not recognized. This came from the fact that the schema of the database was not updated, but only taken from the existing file. Thus, we needed an option to overwrite the existing file.