feat(logging): add dynamic log level from settings
This commit introduces the ability to set the log level dynamically from the settings file. The `get_logger` function now reads the log level from the settings and sets it for the logger. This allows for more flexible control over the logging behavior of the application. By default, if the settings file does not contain the log level, the "INFO" level is set up.