fix: Deal with equal input/output directories in FSLCommand
When input and output directories are equal in FSLCommand, it leads to volumes binding problem in Docker container. Indeed, the 'volumes' are a dictionary indexed on the name of the host directories. Therefore, when they are the same, there will be only one volume.
This commit refactor the FSLCommand abstract base class so that it is taken into account in a higher level.